Completely replace attrs-based config with schematics-based config
