+++ src/csbot/config.py
@@ -1,118 +1,42 @@
-from typing import Type, List, Dict, Any, Optional, Union, TextIO
-import copy
-import os
-import logging
 from contextlib import contextmanager
 from enum import Enum
-import re
+from functools import partial
 import inspect
+import io
+import logging
+import os
+import re
+from typing import Type, TypeVar, List, Dict, Any, Union, Callable, TextIO
 
 import attr
-import cattr
+from schematics import Model, types
+import schematics.exceptions
 import toml
 from toml.encoder import _dump_str
 
+# TODO: add test that checks that the example configuration can be generated and loaded (therefore that no plugins have
+#       required options without an example value)
+# TODO: warn about mutable default/example values
+# TODO: required=True for option_list and option_map?
+# TODO: choices?
+# TODO; custom errors instead of leaking Schematics exceptions?
 
 LOG = logging.getLogger(__name__)
 
 METADATA_KEY = 'csbot_config'
 
-_BOOL_TRUE = {"true", "yes", "1"}
-_BOOL_FALSE = {"false", "no", "0"}
+_TYPE_MAP = {
+    str: types.StringType,
+    int: types.IntType,
+    float: types.FloatType,
+    bool: types.BooleanType,
+}
 
+_T = TypeVar("_T")
 
-def _read_bool(s: str, _cls: Type) -> bool:
-    s_ = s.lower()
-    if s_ in _BOOL_TRUE:
-        return True
-    elif s_ in _BOOL_FALSE:
-        return False
-    else:
-        raise ValueError(f"unrecognised boolean string: {s}")
+_DefaultValue = Union[type(None), _T, Callable[[], _T]]
 
-
-_env_converter = cattr.Converter()
-_env_converter.register_structure_hook(bool, _read_bool)
-
-
-structure = cattr.structure
-unstructure = cattr.unstructure
-
-
-def loads(s, cls):
-    return structure(toml.loads(s), cls)
-
-
-def dumps(o):
-    return toml.dumps(unstructure(o))
-
-
-def load(f, cls):
-    return structure(toml.load(f), cls)
-
-
-def dump(o, f):
-    return toml.dump(unstructure(o), f)
-
-
-class ConfigurationError(Exception):
-    pass
-
-
-class Factory:
-    _example_mode = False
-
-    def __init__(self, cls, *, default: Any = None, example: Any = None, env: List[str] = None):
-        self._cls = cls
-        if callable(default):
-            self._default_callable = default
-        else:
-            self._default_callable = None
-            self._default_value = default
-        if callable(example):
-            self._example_callable = example
-        else:
-            self._example_callable = None
-            self._example_value = example
-        self._env = env or []
-
-    def __call__(self):
-        if self._example_mode:
-            return self._get_example()
-        else:
-            return self._get_default()
-
-    def _get_default(self):
-        # See if any environment variable is populated
-        for var in self._env:
-            if var in os.environ:
-                return _env_converter.structure(os.environ[var], self._cls)
-        # Otherwise use the default
-        if self._default_callable:
-            return self._default_callable()
-        else:
-            return copy.copy(self._default_value)
-
-    def _get_example(self):
-        if self._example_callable:
-            return self._example_callable()
-        elif self._example_value is None:
-            return self._get_default()
-        else:
-            return copy.copy(self._example_value)
-
-    @classmethod
-    @contextmanager
-    def example_mode(cls, enable=True):
-        old = cls._example_mode
-        cls._example_mode = enable
-        yield
-        cls._example_mode = old
-
-
-def make_example(cls):
-    with Factory.example_mode():
-        return cls()
+_example_mode = False
 
 
 class OptionKind(Enum):
@@ -132,107 +56,201 @@ class OptionKind(Enum):
 class OptionMetadata:
     type = attr.ib()
     kind: OptionKind = attr.ib(validator=attr.validators.in_(OptionKind))
-    optional: bool = attr.ib(default=False, validator=attr.validators.instance_of(bool))
     help: str = attr.ib(default="", validator=attr.validators.instance_of(str))
 
 
-def is_structure(cls):
-    return hasattr(cls, '__attrs_attrs__')
+@contextmanager
+def example_mode():
+    global _example_mode
+    old = _example_mode
+    _example_mode = True
+    yield
+    _example_mode = old
+
+
+class Config(Model):
+    def __repr__(self):
+        return f"{self.__class__.__name__}({', '.join(f'{a.name}={repr(a.value)}' for a in self.atoms())})"
+
+
+ConfigError = schematics.exceptions.DataError
+
+
+_SimpleOptionType = Union[Type[str], Type[int], Type[float], Type[bool]]
+_OptionType = Union[Type[Config], _SimpleOptionType]
+
+
+def is_config(obj):
+    if inspect.isclass(obj):
+        return issubclass(obj, Config)
+    else:
+        return isinstance(obj, Config)
 
 
 def is_allowable_type(cls):
-    return is_structure(cls) or cls in (str, bool, int, float)
+    return cls in _TYPE_MAP or is_config(cls)
 
 
-_ATTRS_KWARGS = {
-    "slots": True,
-    "kw_only": True,
-}
+def structure(data: Dict[str, Any], cls: Type[Config]) -> Config:
+    o = cls(data)
+    o.validate()
+    return o
 
 
-def config(cls):
-    return attr.s(**_ATTRS_KWARGS)(cls)
+def unstructure(obj: Config) -> Dict[str, Any]:
+    return obj.to_native()
 
 
-# TODO: "required" options?
-# TODO: user-friendly exceptions (write own validators)
-# TODO: stop strings being structured at list of 1-char strings, raise exception instead
-def option(cls: Type, *, default=None, example=None, env: Union[str, List[str]] = None, help: str) -> attr.Attribute:
-    assert is_allowable_type(cls)
-    type = cls
-    validator = attr.validators.instance_of(cls)
-    if default is None:
-        type = Optional[type]
-        validator = attr.validators.optional(validator)
+def loads(s: str, cls: Type[Config]) -> Config:
+    return structure(toml.loads(s), cls)
+
+
+def dumps(obj: Config) -> str:
+    return toml.dumps(unstructure(obj))
+
+
+def load(f, cls: Type[Config]) -> Config:
+    return structure(toml.load(f), cls)
+
+
+def dump(obj: Config, f):
+    return toml.dump(unstructure(obj), f)
+
+
+class Default:
+    """A callable to get a default or example value.
+
+    Returns *default* (or the result of calling it, if callable) when called.
+
+    If inside a ``with example_mode()``, first tries to return *example* (or the result of calling it, if callable),
+    and falls back to *default* if *example* was None.
+
+    This allows plugin configuration to define required fields that can still generate a useful example without any
+    data being supplied.
+
+    TODO: document *env*
+    """
+    def __init__(self, default: _DefaultValue = None, example: _DefaultValue = None, env: List[str] = None):
+        self._default = default if callable(default) else lambda: default
+        self._example = example if callable(example) else lambda: example
+        self._env = env or []
+
+    def __call__(self):
+        global _example_mode
+        if _example_mode:
+            return self._get_example()
+        else:
+            return self._get_default()
+
+    def _get_default(self, use_env: bool = True):
+        if use_env:
+            for var in self._env:
+                if var in os.environ:
+                    print(f"found {var} in os.environ")
+                    return os.environ[var]
+        return self._default()
+
+    def _get_example(self):
+        example = self._example()
+        if example is None:
+            example = self._get_default(use_env=False)
+        return example
+
+
+def option(cls: _OptionType, *,
+           required: bool = False,
+           default: _DefaultValue = None,
+           example: _DefaultValue = None,
+           env: Union[str, List[str]] = None,
+           help: str) -> types.BaseType:
+    if not is_allowable_type(cls):
+        raise TypeError(f"cls must be subclass of Config or one of {_TYPE_MAP.keys()}")
+
     if isinstance(env, str):
         env = [env]
 
-    attrib_kwargs = {
-        "type": type,
-        "validator": validator,
-        "default": attr.Factory(Factory(cls, default=default, example=example, env=env)),
+    if is_config(cls):
+        field = partial(types.ModelType, cls)
+    else:
+        field = _TYPE_MAP[cls]
+
+    field_kwargs = {
+        "required": required or default is not None,
+        "default": Default(default, example, env),
         "metadata": {
             METADATA_KEY: OptionMetadata(
                 type=cls,
-                kind=OptionKind.STRUCTURE if is_structure(cls) else OptionKind.SIMPLE,
-                optional=default is None,
+                kind=OptionKind.STRUCTURE if is_config(cls) else OptionKind.SIMPLE,
                 help=help,
             ),
         },
     }
+    return field(**field_kwargs)
 
-    return attr.ib(**attrib_kwargs)
 
+def option_list(cls: _OptionType, *,
+                default: _DefaultValue = None,
+                example: _DefaultValue = None,
+                help: str):
+    if not is_allowable_type(cls):
+        raise TypeError(f"cls must be subclass of Config or one of {_TYPE_MAP.keys()}")
 
-def option_list(cls: Type, *, default=None, example=None, help: str) -> attr.Attribute:
-    assert is_allowable_type(cls)
-    type = List[cls]
     if default is None:
         default = list
-    default = attr.Factory(Factory(type, default=default, example=example))
 
-    attrib_kwargs = {
-        "type": type,
-        "validator": attr.validators.deep_iterable(
-            member_validator=attr.validators.instance_of(cls),
-            iterable_validator=attr.validators.instance_of(list),
-        ),
-        "default": default,
+    if is_config(cls):
+        inner_field = types.ModelType(cls, required=True)
+    else:
+        inner_field = _TYPE_MAP[cls](required=True)
+
+    field_kwargs = {
+        "required": True,   # Disallow None as a value, empty list is fine
+        "default": Default(default, example),
         "metadata": {
             METADATA_KEY: OptionMetadata(
                 type=cls,
-                kind=OptionKind.STRUCTURE_LIST if is_structure(cls) else OptionKind.SIMPLE_LIST,
+                kind=OptionKind.STRUCTURE_LIST if is_config(cls) else OptionKind.SIMPLE_LIST,
                 help=help,
             ),
         },
     }
-    return attr.ib(**attrib_kwargs)
+    return types.ListType(inner_field, **field_kwargs)
 
 
-def option_map(cls: Type, *, default=None, example=None, help: str) -> attr.Attribute:
-    assert is_allowable_type(cls)
-    type = Dict[str, cls]
+def option_map(cls: _OptionType, *,
+               default: _DefaultValue = None,
+               example: _DefaultValue = None,
+               help: str):
+    if not is_allowable_type(cls):
+        raise TypeError(f"cls must be subclass of Config or one of {_TYPE_MAP.keys()}")
+
     if default is None:
         default = dict
-    default = attr.Factory(Factory(type, default=default, example=example))
 
-    attrib_kwargs = {
-        "type": type,
-        "validator": attr.validators.deep_mapping(
-            key_validator=attr.validators.instance_of(str),
-            value_validator=attr.validators.instance_of(cls),
-            mapping_validator=attr.validators.instance_of(dict),
-        ),
-        "default": default,
+    if is_config(cls):
+        inner_field = types.ModelType(cls, required=True)
+    else:
+        inner_field = _TYPE_MAP[cls](required=True)
+
+    field_kwargs = {
+        "required": True,   # Disallow None as a value, empty dict is fine
+        "default": Default(default, example),
         "metadata": {
             METADATA_KEY: OptionMetadata(
                 type=cls,
-                kind=OptionKind.STRUCTURE_MAP if is_structure(cls) else OptionKind.SIMPLE_MAP,
+                kind=OptionKind.STRUCTURE_MAP if is_config(cls) else OptionKind.SIMPLE_MAP,
                 help=help,
             ),
         },
     }
-    return attr.ib(**attrib_kwargs)
+    return types.DictType(inner_field, **field_kwargs)
+
+
+def make_example(cls: Type[Config]) -> Config:
+    with example_mode():
+        o = cls()
+        o.validate()
+        return o
 
 
 # TODO: commented output
@@ -279,11 +297,11 @@ class TomlExampleGenerator:
         s = f"{s}\n"
         self._write(s, raw=True)
 
-    def generate(self, obj, stream: TextIO, prefix: List[str] = None):
+    def generate(self, obj: Union[Config, Type[Config]], stream: TextIO, prefix: List[str] = None):
         """Generate an example from *obj* and write it to *stream*."""
         if inspect.isclass(obj):
             obj = make_example(obj)
-        assert is_structure(obj)
+        assert is_config(obj)
         if prefix is None:
             prefix = []
         with self._use_stream(stream):
@@ -291,7 +309,7 @@ class TomlExampleGenerator:
 
     def _generate_option(self,
                          example: Any,
-                         attrib: attr.Attribute,
+                         field: types.BaseType,
                          absolute_path: List[str],
                          relative_path: List[str]):
         """
@@ -304,7 +322,7 @@ class TomlExampleGenerator:
             _generate_structure_list
             _generate_structure_map
         """
-        metadata = self._get_metadata(attrib)
+        metadata = self._get_metadata(field)
         if metadata.help:
             self._writeline(f"## {metadata.help}")
         if metadata.kind is OptionKind.SIMPLE:
@@ -347,7 +365,7 @@ class TomlExampleGenerator:
             for k, v in example.items():
                 self._generate_simple(v, relative_path + [k])
 
-    def _generate_structure(self, example: Any, absolute_path: List[str], is_list_item: bool = False):
+    def _generate_structure(self, example: Config, absolute_path: List[str], is_list_item: bool = False):
         """
         Generate section heading:
             Nothing if top-levell
@@ -369,23 +387,23 @@ class TomlExampleGenerator:
             return
 
         deferred = []
-        for attrib in self._get_attributes(example):
-            metadata = self._get_metadata(attrib)
+        for atom in example.atoms():
+            metadata = self._get_metadata(atom.field)
             if not metadata.kind.is_simple:
                 # Write sections after simple values
-                deferred.append(attrib)
+                deferred.append(atom)
                 continue
-            self._generate_option(getattr(example, attrib.name),
-                                  attrib,
-                                  absolute_path + [attrib.name],
-                                  [attrib.name])
+            self._generate_option(atom.value,
+                                  atom.field,
+                                  absolute_path + [atom.name],
+                                  [atom.name])
 
-        for attrib in deferred:
+        for atom in deferred:
             self._write("\n")
-            self._generate_option(getattr(example, attrib.name),
-                                  attrib,
-                                  absolute_path + [attrib.name],
-                                  [attrib.name])
+            self._generate_option(atom.value,
+                                  atom.field,
+                                  absolute_path + [atom.name],
+                                  [atom.name])
 
     def _generate_structure_list(self, example: List[Any], absolute_path: List[str]):
         """
@@ -418,8 +436,8 @@ class TomlExampleGenerator:
         return getattr(obj, "__attrs_attrs__")
 
     @classmethod
-    def _get_metadata(cls, attrib: attr.Attribute) -> OptionMetadata:
-        return attrib.metadata[METADATA_KEY]
+    def _get_metadata(cls, field: types.BaseType) -> OptionMetadata:
+        return field.metadata[METADATA_KEY]
 
     @classmethod
     def _make_key(cls, path):
@@ -428,8 +446,7 @@ class TomlExampleGenerator:
 
 
 def generate_toml_example(obj, commented=False):
-    from io import StringIO
-    stream = StringIO()
+    stream = io.StringIO()
     generator = TomlExampleGenerator(commented=commented)
     generator.generate(obj, stream)
     return stream.getvalue()
